Struct substrate_keystore::Store

pub struct Store { /* fields omitted */ }

Key store.

Stores key pairs in a file system store + short lived key pairs in memory.

Every pair that is being generated by a seed, will be placed in memory.

Methods

impl Store

pub fn open<T: Into<PathBuf>>(
path: T,
password: Option<Protected<String>>
) -> Result<KeyStorePtr>

Open the store at the given path.

Optionally takes a password that will be used to encrypt/decrypt the keys.

pub fn insert_by_type<Pair: PairT>(
&self,
key_type: KeyTypeId,
suri: &str
) -> Result<Pair>

Insert a new key.

Places it into the file system store.

pub fn insert<Pair: AppPair>(&self, suri: &str) -> Result<Pair>

Insert a new key.

Places it into the file system store.

pub fn generate_by_type<Pair: PairT>(&self, key_type: KeyTypeId) -> Result<Pair>

Generate a new key.

Places it into the file system store.

pub fn generate<Pair: AppPair>(&self) -> Result<Pair>

Generate a new key.

Places it into the file system store.

pub fn insert_ephemeral_from_seed_by_type<Pair: PairT>(
&mut self,
seed: &str,
key_type: KeyTypeId
) -> Result<Pair>

Create a new key from seed.

Does not place it into the file system store.

pub fn insert_ephemeral_from_seed<Pair: AppPair>(
&mut self,
seed: &str
) -> Result<Pair>

Create a new key from seed.

Does not place it into the file system store.

pub fn key_pair_by_type<Pair: PairT>(
&self,
public: &Pair::Public,
key_type: KeyTypeId
) -> Result<Pair>

Get a key pair for the given public key and key type.

pub fn key_pair<Pair: AppPair>(
&self,
public: &<Pair as AppKey>::Public
) -> Result<Pair>

Get a key pair for the given public key.

pub fn public_keys_by_type<TPublic: Public>(
&self,
key_type: KeyTypeId
) -> Result<Vec<TPublic>>

Get public keys of all stored keys that match the given key type.

pub fn public_keys<Public: AppPublic>(&self) -> Result<Vec<Public>>

Get public keys of all stored keys that match the key type.

This will just use the type of the public key (a list of which to be returned) in order to determine the key type. Unless you use a specialised application-type public key, then this only give you keys registered under generic cryptography, and will not return keys registered under the application type.